requirements.txt
pandas
datetime
numpy
progressbar2
configparser
seaborn
matplotlib